refactor: point field
