+++ plugin/web/extensions/sidebar/core/components/content/common/DatasetCard/Field/Fields/point/PointColor.tsx
@@ -1,17 +1,12 @@
 import AddButton from "@web/extensions/sidebar/core/components/content/common/AddButton";
-import { Icon, Dropdown, Menu } from "@web/sharedComponents";
 import { styled } from "@web/theme";
 import { useCallback, useState } from "react";
 
 import { BaseFieldProps, Cond } from "../types";
 
-const operators: { [key: string]: string } = {
-  greater: ">",
-  less: "<",
-  greaterEqual: ">=",
-  lessEqual: "<=",
-  equal: "=",
-};
+import ConditionField from "./ConditionField";
+import Field from "./Field";
+import ItemControls from "./ItemControls";
 
 function array_move(arr: any[], old_index: number, new_index: number) {
   if (new_index >= arr.length) {
@@ -100,30 +95,16 @@ const PointColor: React.FC<BaseFieldProps<"pointColor">> = ({ value, editMode, o
     [value, onUpdate],
   );
 
-  const menu = (
-    <Menu
-      items={Object.keys(operators).map(op => {
-        return {
-          key: op,
-          label: (
-            <p style={{ margin: 0 }} onClick={() => console.log(op)}>
-              {operators[op]}
-            </p>
-          ),
-        };
-      })}
-    />
-  );
-
   return editMode ? (
     <Wrapper>
       {value.pointColors?.map((c, idx) => (
         <Item key={idx}>
-          <ItemControls>
-            <Icon icon="arrowUpThin" size={16} onClick={() => handleMoveUp(idx)} />
-            <Icon icon="arrowDownThin" size={16} onClick={() => handleMoveDown(idx)} />
-            <Icon icon="trash" size={16} onClick={() => handleRemove(idx)} />
-          </ItemControls>
+          <ItemControls
+            index={idx}
+            handleMoveDown={handleMoveDown}
+            handleMoveUp={handleMoveUp}
+            handleRemove={handleRemove}
+          />
           {/* {legend.style === "icon" && (
             <Field>
               <FieldTitle>URL</FieldTitle>
@@ -132,35 +113,17 @@ const PointColor: React.FC<BaseFieldProps<"pointColor">> = ({ value, editMode, o
               </FieldValue>
             </Field>
           )} */}
-          <Field gap={8}>
-            <FieldTitle>if</FieldTitle>
-            <FieldValue>
-              <Dropdown overlay={menu} placement="bottom" trigger={["click"]}>
-                <StyledDropdownButton>
-                  <p style={{ margin: 0 }}>{c.condition.operand}</p>
-                  <Icon icon="arrowDownSimple" size={12} />
-                </StyledDropdownButton>
-              </Dropdown>
-            </FieldValue>
-            <FieldValue>
-              <Dropdown overlay={menu} placement="bottom" trigger={["click"]}>
-                <StyledDropdownButton>
-                  <p style={{ margin: 0 }}>{c.condition.operator}</p>
-                  <Icon icon="arrowDownSimple" size={12} />
-                </StyledDropdownButton>
-              </Dropdown>
-            </FieldValue>
-            <FieldValue>
-              <NumberInput value={c.condition.value} />
-            </FieldValue>
-          </Field>
-          <Field>
-            <FieldTitle width={82}>色</FieldTitle>
-            <FieldValue>
-              <ColorBlock color={c.color} />
-              <TextInput value={c.color} />
-            </FieldValue>
-          </Field>
+          <ConditionField title="if" fieldGap={8} condition={c.condition} />
+          <Field
+            title="色"
+            titleWidth={82}
+            value={
+              <>
+                <ColorBlock color={c.color} />
+                <TextInput value={c.color} />
+              </>
+            }
+          />
         </Item>
       ))}
       <ButtonWrapper>
@@ -188,10 +151,6 @@ const Wrapper = styled.div`
 //   cursor: pointer;
 // `;
 
-const Text = styled.p`
-  margin: 0;
-`;
-
 const Item = styled.div`
   display: flex;
   flex-direction: column;
@@ -201,47 +160,7 @@ const Item = styled.div`
   padding: 8px;
 `;
 
-const ItemControls = styled.div`
-  display: flex;
-  justify-content: right;
-  gap: 4px;
-  cursor: pointer;
-`;
-
-const Field = styled.div<{ gap?: number }>`
-  display: flex;
-  align-items: center;
-  ${({ gap }) => gap && `gap: ${gap}px;`}
-  height: 32px;
-`;
-
-const FieldTitle = styled(Text)<{ width?: number }>`
-  ${({ width }) => width && `width: ${width}px;`}
-`;
-
-const FieldValue = styled.div`
-  display: flex;
-  border: 1px solid #d9d9d9;
-  border-radius: 2px;
-  flex: 1;
-  height: 100%;
-  width: 100%;
-`;
-
 const TextInput = styled.input.attrs({ type: "text" })`
-  height: 100%;
-  width: 100%;
-  flex: 1;
-  padding: 0 12px;
-  border: none;
-  outline: none;
-
-  :focus {
-    border: none;
-  }
-`;
-
-const NumberInput = styled.input.attrs({ type: "number" })`
   height: 100%;
   width: 100%;
   flex: 1;
@@ -276,14 +195,4 @@ const ColorBlock = styled.div<{ color: string; legendStyle?: "circle" | "square"
 const ButtonWrapper = styled.div`
   width: 125px;
   align-self: flex-end;
-`;
-
-const StyledDropdownButton = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  align-content: center;
-  padding: 0 16px;
-  cursor: pointer;
 `;
